# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Cards Management System API
  description: |
    Backend-приложение на Java (Spring Boot) для управления банковскими картами с поддержкой JWT аутентификации.

    ## Функциональность
    - Создание и управление банковскими картами
    - Просмотр карт с маскированием номеров
    - Переводы между своими картами
    - Аутентификация и авторизация с JWT
    - Ролевая модель доступа (ADMIN, USER)

    ## Безопасность
    - Все эндпоинты, кроме /api/auth/*, требуют JWT токен в заголовке Authorization
    - Номера карт хранятся в зашифрованном виде (AES)
    - При отображении номера маскируются: `**** **** **** 1234`
  version: 1.0.0
  contact:
    email: support@bankcards.com
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.bankcards.com
    description: Production server
...
paths:
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Регистрация нового пользователя
      description: Создает нового пользователя с ролью USER
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              username: johndoe
              email: john@example.com
              password: password123
              fullName: John Doe
      responses:
        '200':
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Некорректные данные (username или email уже существует)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Вход в систему
      description: Аутентификация пользователя и получение JWT токена
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: demo
              password: password
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                token: eyJhbGciOiJIUzI1NiJ9...
                type: Bearer
                id: 1
                username: demo
                email: demo@example.com
                fullName: Demo User
                roles:
                  - ROLE_USER
        '401':
          description: Неверные учетные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cards:
    post:
      tags:
        - Cards
      summary: Создание банковской карты
      description: Создает новую банковскую карту для указанного пользователя
      operationId: createCard
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardRequest'
            example:
              cardNumber: "1234567890123456"
              expiryDate: "2027-12-31"
              userId: 1
      responses:
        '201':
          description: Карта успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '400':
          description: Некорректные данные (невалидный номер карты или карта уже существует)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
        '403':
          description: Нет доступа к созданию карты для этого пользователя
  /api/cards/user/{userId}:
    get:
      tags:
        - Cards
      summary: Получение карт пользователя
      description: Возвращает список карт пользователя с пагинацией и опциональной фильтрацией по статусу
      operationId: getUserCards
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID пользователя
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [ACTIVE, BLOCKED, EXPIRED]
          description: Фильтр по статусу карты
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Номер страницы (начиная с 0)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
          description: Количество элементов на странице
      responses:
        '200':
          description: Список карт пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageCardResponse'
        '400':
          description: Некорректный статус карты
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
        '403':
          description: Нет доступа к картам этого пользователя
  /api/cards/{cardId}:
    get:
      tags:
        - Cards
      summary: Получение карты по ID
      description: Возвращает детальную информацию о карте
      operationId: getCardById
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID карты
      responses:
        '200':
          description: Информация о карте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '401':
          description: Не авторизован
        '403':
          description: Нет доступа к этой карте
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Cards
      summary: Удаление карты (только для администраторов)
      description: Удаляет карту из системы
      operationId: deleteCard
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID карты
      responses:
        '204':
          description: Карта успешно удалена
        '401':
          description: Не авторизован
        '403':
          description: Нет прав администратора
        '404':
          description: Карта не найдена
  /api/cards/{cardId}/block:
    put:
      tags:
        - Cards
      summary: Блокировка карты
      description: Блокирует карту (пользователь может заблокировать свою карту)
      operationId: blockCard
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID карты
      responses:
        '200':
          description: Карта успешно заблокирована
        '401':
          description: Не авторизован
        '403':
          description: Нет доступа к этой карте
        '404':
          description: Карта не найдена
  /api/cards/{cardId}/activate:
    put:
      tags:
        - Cards
      summary: Активация карты (только для администраторов)
      description: Активирует заблокированную карту
      operationId: activateCard
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID карты
      responses:
        '200':
          description: Карта успешно активирована
        '401':
          description: Не авторизован
        '403':
          description: Нет прав администратора
        '404':
          description: Карта не найдена
  /api/transfers:
    post:
      tags:
        - Transfers
      summary: Создание перевода
      description: Создает перевод между двумя картами пользователя
      operationId: createTransfer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
            example:
              fromCardId: 1
              toCardId: 2
              amount: 100.00
              description: Transfer between my cards
      responses:
        '201':
          description: Перевод успешно выполнен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResponse'
        '400':
          description: Некорректные данные (недостаточно средств, карты заблокированы и т.д.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
        '403':
          description: Нет доступа к указанным картам
        '404':
          description: Одна из карт не найдена
  /api/transfers/card/{cardId}:
    get:
      tags:
        - Transfers
      summary: Получение истории переводов карты
      description: Возвращает историю переводов для указанной карты с пагинацией
      operationId: getCardTransferHistory
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID карты
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Номер страницы
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
          description: Количество элементов на странице
      responses:
        '200':
          description: Список переводов карты
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageTransferResponse'
        '401':
          description: Не авторизован
        '403':
          description: Нет доступа к этой карте
        '404':
          description: Карта не найдена
  /api/transfers/user/{userId}:
    get:
      tags:
        - Transfers
      summary: Получение истории переводов пользователя
      description: Возвращает всю историю переводов для пользователя с пагинацией
      operationId: getUserTransferHistory
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID пользователя
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Номер страницы
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
          description: Количество элементов на странице
      responses:
        '200':
          description: Список переводов пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageTransferResponse'
        '401':
          description: Не авторизован
        '403':
          description: Нет доступа к переводам этого пользователя
  /api/transfers/{transferId}:
    get:
      tags:
        - Transfers
      summary: Получение информации о переводе по ID
      description: Возвращает детальную информацию о конкретном переводе
      operationId: getTransferById
      security:
        - bearerAuth: []
      parameters:
        - name: transferId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID перевода
      responses:
        '200':
          description: Информация о переводе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResponse'
        '401':
          description: Не авторизован
        '403':
          description: Нет доступа к этому переводу
        '404':
          description: Перевод не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен, полученный при login/register
  schemas:
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
        - fullName
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          description: Имя пользователя (уникальное)
          example: johndoe
        email:
          type: string
          format: email
          maxLength: 100
          description: Email адрес (уникальный)
          example: john@example.com
        password:
          type: string
          minLength: 6
          maxLength: 100
          description: Пароль (минимум 6 символов)
          example: password123
        fullName:
          type: string
          minLength: 1
          maxLength: 100
          description: Полное имя пользователя
          example: John Doe
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: Имя пользователя
          example: demo
        password:
          type: string
          format: password
          description: Пароль
          example: password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT токен
          example: eyJhbGciOiJIUzI1NiJ9...
        type:
          type: string
          description: Тип токена
          example: Bearer
          default: Bearer
        id:
          type: integer
          format: int64
          description: ID пользователя
          example: 1
        username:
          type: string
          description: Имя пользователя
          example: johndoe
        email:
          type: string
          format: email
          description: Email пользователя
          example: john@example.com
        fullName:
          type: string
          description: Полное имя
          example: John Doe
        roles:
          type: array
          items:
            type: string
          description: Роли пользователя
          example: [ROLE_USER]
    CardRequest:
      type: object
      required:
        - cardNumber
        - expiryDate
        - userId
      properties:
        cardNumber:
          type: string
          pattern: '^\d{16}$'
          description: Номер карты (16 цифр)
          example: "1234567890123456"
        expiryDate:
          type: string
          format: date
          description: Дата истечения срока действия
          example: "2027-12-31"
        userId:
          type: integer
          format: int64
          description: ID владельца карты
          example: 1
    CardResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID карты
          example: 1
        cardNumber:
          type: string
          description: Маскированный номер карты
          example: "**** **** **** 1234"
        expiryDate:
          type: string
          format: date
          description: Дата истечения срока действия
          example: "2027-12-31"
        status:
          type: string
          enum: [ACTIVE, BLOCKED, EXPIRED]
          description: Статус карты
          example: ACTIVE
        balance:
          type: number
          format: decimal
          description: Баланс карты
          example: 1000.00
        userId:
          type: integer
          format: int64
          description: ID владельца
          example: 1
        createdAt:
          type: string
          format: date-time
          description: Дата создания
          example: "2025-01-15T10:30:00"
        updatedAt:
          type: string
          format: date-time
          description: Дата последнего обновления
          example: "2025-01-15T10:30:00"
    TransferRequest:
      type: object
      required:
        - fromCardId
        - toCardId
        - amount
      properties:
        fromCardId:
          type: integer
          format: int64
          description: ID карты отправителя
          example: 1
        toCardId:
          type: integer
          format: int64
          description: ID карты получателя
          example: 2
        amount:
          type: number
          format: decimal
          minimum: 0.01
          description: Сумма перевода
          example: 100.00
        description:
          type: string
          maxLength: 500
          description: Описание перевода
          example: Transfer between my cards
    TransferResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID перевода
          example: 1
        fromCardId:
          type: integer
          format: int64
          description: ID карты отправителя
          example: 1
        toCardId:
          type: integer
          format: int64
          description: ID карты получателя
          example: 2
        amount:
          type: number
          format: decimal
          description: Сумма перевода
          example: 100.00
        description:
          type: string
          description: Описание перевода
          example: Transfer between my cards
        status:
          type: string
          enum: [PENDING, COMPLETED, FAILED]
          description: Статус перевода
          example: COMPLETED
        createdAt:
          type: string
          format: date-time
          description: Дата создания
          example: "2025-01-15T10:30:00"
...
    PageCardResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CardResponse'
        pageable:
          $ref: '#/components/schemas/Pageable'
        totalElements:
          type: integer
          format: int64
          description: Общее количество элементов
          example: 100
        totalPages:
          type: integer
          description: Общее количество страниц
          example: 10
        size:
          type: integer
          description: Размер страницы
          example: 10
        number:
          type: integer
          description: Номер текущей страницы
          example: 0
        first:
          type: boolean
          description: Первая страница?
          example: true
        last:
          type: boolean
          description: Последняя страница?
          example: false
        empty:
          type: boolean
          description: Пустая страница?
          example: false
    PageTransferResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TransferResponse'
        pageable:
          $ref: '#/components/schemas/Pageable'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          description: Номер страницы
          example: 0
        size:
          type: integer
          description: Размер страницы
          example: 10
        sort:
          type: array
          items:
            type: string
          description: Параметры сортировки
          example: ["createdAt,desc"]
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки
          example: "2025-01-15T10:30:00"
        status:
          type: integer
          description: HTTP статус код
          example: 400
        error:
          type: string
          description: Описание ошибки
          example: Bad Request
        message:
          type: string
          description: Детальное сообщение об ошибке
          example: Card number already exists
        path:
          type: string
          description: Путь запроса
          example: /api/cards