# src/main/java/com/example/bankcards/dto/CardResponse.java
package com.example.bankcards.dto;

import java.math.BigDecimal;
import java.time.LocalDate;

public record CardResponse(
        Long id,
        String cardNumber,
        String ownerFullName,
        LocalDate expiryDate,
        String status,
        BigDecimal balance
) {}