# src/main/java/com/example/bankcards/dto/ErrorResponse.java
package com.example.bankcards.dto;

import java.time.LocalDateTime;
import java.util.Map;

public record ErrorResponse(
        LocalDateTime timestamp,
        int status,
        String error,
        String message,
        String path,
        Map<String, String> validationErrors
) {}